# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a server for work with metrics",
        "title": "SystemMetric API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/deleteAlert": {
            "delete": {
                "description": "delete alert",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "API"
                ],
                "summary": "delete page",
                "parameters": [
                    {
                        "type": "string",
                        "description": "alertID for deleting alert",
                        "name": "alertID",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "delete alert is successful",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/deleteAlertRecipient": {
            "delete": {
                "description": "delete alert recipient with alertRecipientID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "API"
                ],
                "summary": "delete page",
                "parameters": [
                    {
                        "type": "string",
                        "description": "alertRecipientID for deleting alert recipient",
                        "name": "alertRecipientID",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "delete alert recipient is successful",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/getAlert": {
            "get": {
                "description": "get alert with alertID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "API"
                ],
                "summary": "get page",
                "parameters": [
                    {
                        "type": "string",
                        "description": "alertID for getting alert",
                        "name": "alertID",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get alert is successful",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/getAlertRecipient": {
            "get": {
                "description": "get alert recipient with alertRecipientID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "API"
                ],
                "summary": "get page",
                "parameters": [
                    {
                        "type": "string",
                        "description": "alertRecipientID for getting alert recipient",
                        "name": "alertRecipientID",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get alert recipient is successful",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/getAlertRecipients": {
            "get": {
                "description": "get alert recipients",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "API"
                ],
                "summary": "get page",
                "responses": {
                    "200": {
                        "description": "Get alert recipients is successful",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/getAlerts": {
            "get": {
                "description": "get alerts",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "API"
                ],
                "summary": "get page",
                "responses": {
                    "200": {
                        "description": "Get alerts is successful",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/insertAlert": {
            "post": {
                "description": "insert alert",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "API"
                ],
                "summary": "insert page",
                "parameters": [
                    {
                        "description": "alert_message, severity, metric_id",
                        "name": "alert",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.Alert"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Insert alert is successful",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/insertAlertRecipient": {
            "post": {
                "description": "insert alert recipient",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "API"
                ],
                "summary": "insert page",
                "parameters": [
                    {
                        "description": "alert_id, user_id",
                        "name": "alert-recipient",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.AlertRecipient"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Insert alert recipient is successful",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "entity.Alert": {
            "type": "object",
            "properties": {
                "alert_id": {
                    "type": "integer",
                    "example": 1
                },
                "alert_message": {
                    "type": "string",
                    "example": "alert message"
                },
                "metric_id": {
                    "type": "integer",
                    "example": 1
                },
                "severity": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "entity.AlertRecipient": {
            "type": "object",
            "properties": {
                "alert_id": {
                    "type": "integer",
                    "example": 1
                },
                "alert_recipient_id": {
                    "type": "integer",
                    "example": 1
                },
                "user_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
    }
}